airflow/requirements.txt
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
alembic==1.4.2
asgiref==3.5.0
async-timeout==4.0.2
asyncpg==0.25.0
asyncpgsa==0.27.1
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-scheduler==0.9.5
frozenlist==1.3.0
greenlet==1.1.2
icalendar==4.0.9
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.1.0
multidict==6.0.2
psycopg2-binary==2.8.5
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.8
yarl==1.7.2
zipp==3.7.0
